refactor: restructure modules folder to follow conventions
