+++ src/app/(school)/(reviews)/@header/page.tsx
@@ -1,4 +1,4 @@
-import { AnnouncementCarousel } from "@/modules/home/AnnouncementCarousel";
+import { AnnouncementCarousel } from "@/modules/home/components/AnnouncementCarousel";
 import { getFeatureFlag } from "@/server/posthog";
 
 export default async function HomeHeader() {
