+++ src/app/(school)/(reviews)/@information/course/[code]/page.tsx
@@ -1,7 +1,7 @@
 import { api } from "@/common/tools/trpc/server";
 import { notFound } from "next/navigation";
-import { InformationCard } from "@/modules/reviews/InformationSection/InformationCard";
-import { DetailCard } from "@/modules/reviews/InformationSection/DetailCard";
+import { InformationCard } from "@/modules/reviews/components/InformationSection/InformationCard";
+import { DetailCard } from "@/modules/reviews/components/InformationSection/DetailCard";
 import { getServerAuthSession } from "@/server/auth";
 
 export default async function CourseInfo({
