+++ src/modules/home/components/AnnouncementBanner/index.ts
