+++ src/modules/home/components/AnnouncementCarousel/AnnouncementCard/AnnouncementCard.tsx
