+++ src/modules/home/components/AnnouncementCarousel/AnnouncementCard/index.ts
