+++ src/modules/home/components/AnnouncementCarousel/index.ts
