+++ src/modules/reviews/components/InformationSection/DetailCard/DetailCard.theme.ts
