+++ src/modules/reviews/components/InformationSection/DetailCard/DetailCard.tsx
