+++ src/modules/reviews/components/InformationSection/DetailCard/DetailCardSkeleton.tsx
