+++ src/modules/reviews/components/InformationSection/DetailCard/index.ts
