+++ src/modules/reviews/components/InformationSection/InformationCard/InformationCard.theme.ts
