+++ src/modules/reviews/components/InformationSection/InformationCard/InformationCard.tsx
