+++ src/modules/reviews/components/InformationSection/InformationCard/InformationCardSkeleton.tsx
