+++ src/modules/reviews/components/InformationSection/InformationCard/InformationModal.tsx
