+++ src/modules/reviews/components/InformationSection/InformationCard/index.ts
