+++ src/modules/submit/constants/index.ts
@@ -1,0 +1,1 @@
+export * from "./texts";
