+++ src/modules/submit/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./reviewable";
+export * from "./reviewer";
