+++ src/modules/submit/types/reviewer.ts
@@ -1,0 +1,4 @@
+export enum ReviewerEnum {
+  ANONYMOUS = "anonymous",
+  USER = "user",
+}
